+++ NEWS.md
@@ -14,6 +14,8 @@
 
 - `check_old_package()` has been vectorized, i.e., the arguments `name` and `version` can take vectors now.
 
+- Factored out the code for parsing chunk options and dividing a chunk into chunk options and chunk body from **knitr** to this package as functions `csv_options()` and `divide_chunk()`, respectively. They will be used by **knitr** and other packages in future.
+
 # CHANGES IN xfun VERSION 0.41
 
 - `process_file()` will write to the file only if the processed text is different with the input text. This is to avoid writing files unnecessarily.